designate text parsing tests.
